test setup.sh

+++ setup.sh
@@ -302,8 +302,10 @@ function start_recursive {
     # 8 - Install
     ! $dep_recursive && { info_print "\n (8) Start compiling"; }
     # Copy headers in main lib folder
-    cp "$location"/res/include/* "$ud_lib_path"/include/
-    is_error $? && { error_print "Copy headers files to [ $ud_lib_path/include/ ] failed"; }
+    if [ -f "$ud_lib_path"/include/ud_"$target_name".h ] || [ diff "$location"/res/include/ud_"$target_name".h "$ud_lib_path"/include/ud_"$target_name".h != "" ] ; then
+        cp "$location"/res/include/* "$ud_lib_path"/include/
+        is_error $? && { error_print "Copy headers files to [ $ud_lib_path/include/ ] failed"; }
+    fi
     # Compil
     if ! $dep_recursive ; then
         make -C "$location" --no-print- LIBNAME="$target_name" DEPNAME="$make_dep_name" ARNAME="$make_ar_name" DEPHEADER="$dep_header" >&2
